Move central schema validation into the task that builds it
