+++ gradle-plugins/application-plugin/src/main/kotlin/viaduct/gradle/ViaductApplicationPlugin.kt
@@ -2,7 +2,6 @@ package viaduct.gradle
 
 import centralSchemaDirectory
 import grtClassesDirectory
-import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.attributes.Category
@@ -13,7 +12,6 @@ import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.api.tasks.bundling.Jar
 import org.gradle.kotlin.dsl.register
-import org.slf4j.LoggerFactory
 import viaduct.gradle.ViaductPluginCommon.addViaductDependencies
 import viaduct.gradle.ViaductPluginCommon.addViaductTestFixtures
 import viaduct.gradle.ViaductPluginCommon.applyViaductBOM
@@ -43,32 +41,7 @@ class ViaductApplicationPlugin : Plugin<Project> {
             val assembleCentralSchemaTask = setupAssembleCentralSchemaTask()
             setupOutgoingConfigurationForCentralSchema(assembleCentralSchemaTask)
 
-            // Create validate schema task at project level
-            val validateSchemaTask = tasks.register("validateViaductSchema") {
-                group = "viaduct"
-                description = "Validate the central GraphQL schema. Fails the build if invalid."
-                dependsOn(assembleCentralSchemaTask)
-
-                // Capture schema directory as a provider to avoid Project reference in doLast
-                val schemaDir = centralSchemaDirectory()
-                val schemaFiles = schemaDir.map { it.asFileTree.matching { include("**/*.graphqls") } }
-
-                doLast {
-                    // Get schema files at execution time instead of using inputs
-                    val filesToValidate = schemaFiles.get().files.toList()
-                    val logger = LoggerFactory.getLogger(ViaductApplicationPlugin::class.java)
-                    val validator = ViaductBasicSchemaValidator(logger)
-                    val errors = validator.validateSchema(filesToValidate)
-                    if (errors.isNotEmpty()) {
-                        errors.forEach { logger.error(it.message ?: it.toString()) }
-                        throw GradleException("GraphQL schema validation failed. See errors above.")
-                    } else {
-                        logger.info("GraphQL schema validation successful.")
-                    }
-                }
-            }
-
-            val generateGRTsTask = setupGenerateGRTsTask(appExt, assembleCentralSchemaTask, validateSchemaTask)
+            val generateGRTsTask = setupGenerateGRTsTask(appExt, assembleCentralSchemaTask)
 
             setupConsumableConfigurationForGRT(generateGRTsTask.flatMap { it.archiveFile })
 
@@ -95,12 +68,10 @@ class ViaductApplicationPlugin : Plugin<Project> {
     private fun Project.setupGenerateGRTsTask(
         appExt: ViaductApplicationExtension,
         assembleCentralSchemaTask: TaskProvider<AssembleCentralSchemaTask>,
-        validateSchemaTask: TaskProvider<*>,
     ): TaskProvider<Jar> {
         val pluginClasspath = files(ViaductPluginCommon.getClassPathElements(this@ViaductApplicationPlugin::class.java))
 
         val generateGRTClassesTask = tasks.register<GenerateGRTClassFilesTask>("generateViaductGRTClassFiles") {
-            dependsOn(validateSchemaTask)
             grtClassesDirectory.set(grtClassesDirectory())
             schemaFiles.setFrom(assembleCentralSchemaTask.flatMap { it.outputDirectory.map { dir -> dir.asFileTree.matching { include("**/*.graphqls") }.files } })
             grtPackageName.set(appExt.grtPackageName)
@@ -144,14 +115,18 @@ class ViaductApplicationPlugin : Plugin<Project> {
 
     private fun Project.setupConsumableConfigurationForGRT(artifact: Provider<RegularFile>) {
         configurations.create(ViaductPluginCommon.Configs.GRT_CLASSES_OUTGOING).apply {
-            description = "Consumable configuration for the jar file containing the GRT classes plus the central schema's graphqls file."
+            description =
+                "Consumable configuration for the jar file containing the GRT classes plus the central schema's graphqls file."
             isCanBeConsumed = true
             isCanBeResolved = false
             attributes {
                 attribute(ViaductPluginCommon.VIADUCT_KIND, ViaductPluginCommon.Kind.GRT_CLASSES)
                 attribute(Usage.USAGE_ATTRIBUTE, objects.named(Usage::class.java, Usage.JAVA_RUNTIME))
                 attribute(Category.CATEGORY_ATTRIBUTE, objects.named(Category::class.java, Category.LIBRARY))
-                attribute(LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE, objects.named(LibraryElements::class.java, LibraryElements.JAR))
+                attribute(
+                    LibraryElements.LIBRARY_ELEMENTS_ATTRIBUTE,
+                    objects.named(LibraryElements::class.java, LibraryElements.JAR)
+                )
             }
             outgoing.artifact(artifact)
         }
